fix(venner-core): use u32 for timestamps, add widget state tests
- Change updated_at, last_updated_at, exported_at from u64 to u32 for portability
- Add reduce_handles_widget_state_patch_and_transient test
- Add export_import_roundtrip_keeps_new_widget_fields test

+++ crates/venner_core/src/state/app_state.rs
@@ -127,7 +127,7 @@ pub struct WidgetState {
     #[serde(default)]
     pub transient: serde_json::Value,
     #[serde(default)]
-    pub updated_at: u64,
+    pub updated_at: u32,
     #[serde(default)]
     pub source: String,
 }
@@ -163,7 +163,7 @@ pub struct UiState {
     #[serde(default)]
     pub panels: HashMap<String, bool>,
     #[serde(default)]
-    pub last_updated_at: u64,
+    pub last_updated_at: u32,
 }
 
 impl Default for UiState {
@@ -202,7 +202,7 @@ impl Default for ThemeTokens {
 #[serde(rename_all = "camelCase")]
 pub struct PersistedAppState {
     pub schema_version: u32,
-    pub exported_at: u64,
+    pub exported_at: u32,
     pub app_version: String,
     pub state: AppState,
 }
